requirements.txt
httpx
beautifulsoup4
lxml
fastapi
uvicorn
pydantic
sentence-transformers
faiss-cpu
anthropic
